+++ source/dotnet/Services/Infrastructure/SettlementReports/SettlementReportRepository.cs
@@ -34,9 +34,9 @@ public class SettlementReportRepository : ISettlementReportRepository
         _streamZipper = streamZipper;
         _fileIdentifierProviders = new List<Func<Guid, GridAreaCode, (string Directory, string Extension, string EntryPath)>>
         {
-            GetTimeSeriesHourBasisDataFileSpecification,
-            GetTimeSeriesQuarterBasisDataFileSpecification,
-            GetMasterBasisDataFileSpecification,
+            GetTimeSeriesHourBasisDataForTotalGridAreaFileSpecification,
+            GetTimeSeriesQuarterBasisDataForTotalGridAreaFileSpecification,
+            GetMasterBasisDataFileForTotalGridAreaSpecification,
         };
     }
 
@@ -58,18 +58,36 @@ public class SettlementReportRepository : ISettlementReportRepository
         return new SettlementReport(stream);
     }
 
-    public static (string Directory, string Extension, string ZipEntryPath) GetTimeSeriesHourBasisDataFileSpecification(Guid batchId, GridAreaCode gridAreaCode)
+    public static (string Directory, string Extension, string ZipEntryPath) GetTimeSeriesHourBasisDataForTotalGridAreaFileSpecification(Guid batchId, GridAreaCode gridAreaCode)
         => ($"calculation-output/batch_id={batchId}/basis_data/time_series_hour/grouping=total_ga/grid_area={gridAreaCode.Code}/",
             ".csv",
             $"{gridAreaCode.Code}/Timeseries_PT1H.csv");
 
-    public static (string Directory, string Extension, string ZipEntryPath) GetTimeSeriesQuarterBasisDataFileSpecification(Guid batchId, GridAreaCode gridAreaCode)
+    public static (string Directory, string Extension, string ZipEntryPath) GetTimeSeriesQuarterBasisDataForTotalGridAreaFileSpecification(Guid batchId, GridAreaCode gridAreaCode)
         => ($"calculation-output/batch_id={batchId}/basis_data/time_series_quarter/grouping=total_ga/grid_area={gridAreaCode.Code}/",
             ".csv",
             $"{gridAreaCode.Code}/Timeseries_PT15M.csv");
 
-    public static (string Directory, string Extension, string ZipEntryPath) GetMasterBasisDataFileSpecification(Guid batchId, GridAreaCode gridAreaCode)
+    public static (string Directory, string Extension, string ZipEntryPath) GetMasterBasisDataFileForTotalGridAreaSpecification(Guid batchId, GridAreaCode gridAreaCode)
         => ($"calculation-output/batch_id={batchId}/basis_data/master_basis_data/grouping=total_ga/grid_area={gridAreaCode.Code}/",
+            ".csv",
+            $"{gridAreaCode.Code}/MeteringPointMasterData.csv");
+
+    public static (string Directory, string Extension, string ZipEntryPath)
+        GetTimeSeriesHourBasisDataForEsPerGaGridAreaFileSpecification(Guid batchId, GridAreaCode gridAreaCode, string energySupplierId)
+        => ($"calculation-output/batch_id={batchId}/basis_data/time_series_hour/grouping=es_ga/grid_area={gridAreaCode.Code}/energy_supplier_gln={energySupplierId}/",
+            ".csv",
+            $"{gridAreaCode.Code}/Timeseries_PT1H.csv");
+
+    public static (string Directory, string Extension, string ZipEntryPath)
+        GetTimeSeriesQuarterBasisDataForEsPerGaFileSpecification(Guid batchId, GridAreaCode gridAreaCode, string energySupplierId)
+        => ($"calculation-output/batch_id={batchId}/basis_data/time_series_quarter/grouping=es_ga/grid_area={gridAreaCode.Code}/energy_supplier_gln={energySupplierId}/",
+            ".csv",
+            $"{gridAreaCode.Code}/Timeseries_PT15M.csv");
+
+    public static (string Directory, string Extension, string ZipEntryPath)
+        GetMasterBasisDataFileForForEsPerGaSpecification(Guid batchId, GridAreaCode gridAreaCode, string energySupplierId)
+        => ($"calculation-output/batch_id={batchId}/basis_data/master_basis_data/grouping=es_ga/grid_area={gridAreaCode.Code}/energy_supplier_gln={energySupplierId}/",
             ".csv",
             $"{gridAreaCode.Code}/MeteringPointMasterData.csv");